[IR] Move IrSignaturesExtractor to a common compiler module
Extract `IrSignaturesExtractor` from a KLIB tool-module to
a common IR serialization module to make it possible to use
the component in Native backend.
